Meter values can be float

+++ MyFlat/Dto/MeterChildDto.cs
@@ -4,7 +4,7 @@
     {
         public int Id_counter { get; set; }
         public string Nm_counter { get; set; }
-        public int Vl_last_indication { get; set; }
+        public float Vl_last_indication { get; set; }
         public string Dt_last_indication { get; set; }
     }
 }
